Using Berkshelf for the Cookbooks. Make Nginx web server working.
From now on the cookbooks will not stay at this place, because Berkshelf [1] will take care of it.

You have to install Berkshelf on you system and this should all what you have to do. Read the documentation on their website for further informations.

Under the hood it does nearly the same way as before (nginx support, disable modules, database cookbook and language support) but does some of this in a different way. Especially the nginx support was refactored. Additional the drupal cookbook was moved to a dedicated github repository [2] or better it is a modified fork. Have a look there for more informations.

[1] http://berkshelf.com/
[2] https://github.com/Konafets/drupal-cookbook